bugfix: fixes (use <unit>) crashes
before this patch, (use posix), for example, would result in a segfault.

Note that you need to recompile and redeploy all your eggs after
recompiling your android-chicken-install with this patch, otherwise your
eggs will suffer the same segfault as your units.

Steps to reproduce:

    adb push android-chicken/build/target/ /
    adb shell csi
    #> (use posix)
    ... segfault

    adb push android-chicken/build/target/ /
    adb shell csi
    #> (use posix)
    #> directory?
    #<procedure>

The fix for the problem was discovered by Felix.